Add simple drafts page

+++ app/api/record/[id]/route.js
@@ -4,8 +4,15 @@ import { getServerSession } from 'next-auth';
 import lib from '../../../../lib/lib'
 
 export async function GET(request, { params }) {
-
     const session = await getServerSession(authOptions);
+    if (!session) {
+        return Response.json({
+            'status': 'error',
+            'message': 'Not authorized'
+        }, {
+            status: 401
+        })
+    }
 
     let where = lib.limitQueryByFamily({ id: params.id }, request.cookies, session)
     const record = await prisma.record.findFirst({
@@ -29,6 +36,14 @@ export async function GET(request, { params }) {
 
 export async function PUT(request, { params }) {
     const session = await getServerSession(authOptions)
+    if (!session) {
+        return Response.json({
+            'status': 'error',
+            'message': 'Not authorized'
+        }, {
+            status: 401
+        })
+    }
 
     let formData = await request.formData()
     formData = Object.fromEntries(formData)
@@ -77,6 +92,14 @@ export async function PUT(request, { params }) {
 
 export async function DELETE(request, { params }) {
     const session = await getServerSession(authOptions);
+    if (!session) {
+        return Response.json({
+            'status': 'error',
+            'message': 'Not authorized'
+        }, {
+            status: 401
+        })
+    }
 
     let where = lib.limitQueryByFamily({ id: params.id }, request.cookies, session)
     const record = await prisma.record.deleteMany({
